ci(workflows): pin action versions
This commit uses the standard convention of specifying the versions
of GitHub actions in a semver comment after the SHA.
Tools like renovate will use this, and pick it up.

It will lead to better PR's from renovate, where instead of
getting PR's like
Currently diff: 2342342342 -> 344939345328
With this: v4.2.3 -> v4.2.4

Dont worry though, the SHA-hash is still connected to
the version. If not, Renovate will adjust/correct this.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>