ci(commit): extend description length
The default setting for the lint committer
(conform) is 72 for description length.

This will at times leads to fails, as for
example tools like Dependabot and Renovate
sometimes adds longer commits messages.
Bumping it to 92 in length, has worked fine
in other projects, so should be fine here to
even for those longer cases.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>