feat: trim ci (renovate, openssf) (#431)
* ci(ssfscorecard): adjust scorecard settings

This PR trims the open ssf scorecard settings so
that it

A) runs twice a week only
B) is runnable from GUI anytime

A) As it is more of general health tool,
there is no need to run it every commit to main.

B) Leaves the option of being able to run it anytime

Signed-off-by: Josef Andersson <josef.andersson@digg.se>

* ci(renovate): use base renovate config

This PR trims and aligns the renovate config
to use the base organisation config.
It eases the maintenance of administration,
but still, it also makes sense when juggling many repositories.

100% compatible, i.e no functionality loss to current conf.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>

* ci(reuseableci): pin sha and version

Signed-off-by: Josef Andersson <josef.andersson@digg.se>

---------

Signed-off-by: Josef Andersson <josef.andersson@digg.se>